Added Diff polling speed setting
Switched number input for settings to string (needed evil). This avoids premature clamping and prevents issues with partial or invalid number input (e.g., typing -, 5., or empty).

Changed trackDiffFrames() because the interval was made, but couldn't be changed, so we need a reactive element, I used a reactive effect (e.g. createEffect) to watch for changes. After which we clear and reset the interval whenever diffLogPolling() changes in the Store